// src/app/enums/news/SpeechPartEnum.ts
export enum SpeechPartEnum {
  Undefined = 0,
  Noun = 1,
  Verb = 2,
  Adjective = 3,
  Adverb = 4,
  Conjunction = 5,
  Determiner = 6,
  Number = 7,
  OrdinalNumber = 8,
  Preposition = 9,
  Predeterminer = 10,
  Pronoun = 11,
  Prefix = 12,
  Suffix = 13,
  Exclamation = 14,
  Idiom = 15,
  Phrase = 16,
}